\docType{package}
\name{sgtools}
\alias{package-sgtools}
\alias{sgtools}
\alias{sgtools-package}
\title{miscellaneous R functions}
\description{
  This package provides some miscellaneous functions.
}
\details{
  \tabular{ll}{ Package: \tab sgtools \cr Version: \tab 0.1
  \cr Date: \tab 2012-07-21 \cr License: \tab GPL (>= 3)\cr
  URL: \tab http://www.github.com/sgibb/sgtools/ \cr }
}
\author{
  Sebastian Gibb <\email{mail@sebastiangibb.de}>
}
\references{
  \url{http://www.github.com/sgibb/sgtools/}
}
\keyword{package}

