\name{zoomPlot}
\alias{zoomPlot}
\title{zoomPlot}
\usage{
  zoomPlot()
}
\description{
  zoom into a user defined region (spanned by two clicks)
  \emph{double click on bottom left edge}: return to
  original size \emph{double click on top right edge}:
  return to last zoom level \emph{right click}: quit
}

